Extracted new wikiService for SoCraTes to help keep code and tests more concise.
