+++ socrates/lib/wiki/index.js
@@ -5,7 +5,7 @@ var Renderer = beans.get('renderer');
 var wikiService = beans.get('wikiService');
 var statusmessage = beans.get('statusmessage');
 var misc = beans.get('misc');
-var socratesActivitiesService = beans.get('socratesActivitiesService');
+var socratesWikiService = beans.get('socratesWikiService');
 var currentYear = beans.get('socratesConstants').currentYear;
 
 function showPage(subdir, pageName, pageVersion, req, res, next) {
@@ -106,7 +106,7 @@ app.post('/rename/:subdir/:page', function (req, res, next) {
 
 app.get('/:year/participantsOverview/', function (req, res, next) {
   var year = req.params.year;
-  socratesActivitiesService.getParticipantsFor(year, function (err, participants) {
+  socratesWikiService.getParticipantsFor(year, function (err, participants) {
     if (err || !participants) { return next(err); }
     res.render('participants', {
       title: 'Participants',
