+++ socrates/test/wiki/wikiService_test.js
@@ -1,0 +1,101 @@
+/* eslint no-underscore-dangle: 0 */
+'use strict';
+
+var sinon = require('sinon').sandbox.create();
+var expect = require('must-dist');
+var _ = require('lodash');
+
+var beans = require('../../testutil/configureForTest').get('beans');
+
+var socratesWikiService = beans.get('socratesWikiService');
+var activitystore = beans.get('activitystore');
+var SoCraTesActivity = beans.get('socratesActivity');
+var Member = beans.get('member');
+var Subscriber = beans.get('subscriber');
+
+var memberstore = beans.get('memberstore');
+var subscriberstore = beans.get('subscriberstore');
+
+var events = beans.get('events');
+var eventstore = beans.get('eventstore');
+var GlobalEventStore = beans.get('GlobalEventStore');
+
+describe('SoCraTes Wiki Service', function () {
+
+  var eventStore;
+  var subscriber;
+
+  beforeEach(function () {
+    eventStore = new GlobalEventStore();
+
+    sinon.stub(eventstore, 'getEventStore', function (url, callback) {
+      callback(null, eventStore);
+    });
+
+    sinon.stub(memberstore, 'getMembersForIds', function (ids, callback) {
+      callback(null, _.map(ids, function (id) { return new Member({id: id}); }));
+    });
+
+    sinon.stub(subscriberstore, 'allSubscribers', function (callback) {
+      callback(null, [subscriber]);
+    });
+  });
+
+  afterEach(function () {
+    sinon.restore();
+  });
+
+  it('loads the participants and their participation information for a year (before 2016)', function (done) {
+    /*eslint camelcase: 0*/
+    var socrates = {
+      id: 'socratesId',
+      title: 'SoCraTes',
+      description: 'Coolest event ever :-)',
+      location: 'Right next door',
+      url: 'socrates-url',
+      isSoCraTes: true,
+      startUnix: 1440687600,
+      endUnix: 1440946800,
+      owner: {nickname: 'ownerNick'},
+      assignedGroup: 'assignedGroup',
+      group: {groupLongName: 'longName'},
+      resources: {
+        single: {
+          _canUnsubscribe: false,
+          _limit: 10,
+          _registrationOpen: true,
+          _registeredMembers: [{memberId: 'memberId', duration: 2}]
+        },
+        bed_in_double: {_canUnsubscribe: false, _limit: 10, _registrationOpen: true}
+      }
+    };
+    sinon.stub(activitystore, 'getActivity', function (url, callback) {
+      callback(null, new SoCraTesActivity(socrates));
+    });
+
+    subscriber = new Subscriber({id: 'memberId'});
+    subscriber.participationOf('2010').state.roommate = 'My buddy';
+
+    socratesWikiService.getParticipantsFor('2010', function (err, participants) {
+      expect(participants).to.have.length(1);
+      expect(participants[0].participation.roommate()).to.be('My buddy');
+      done(err);
+    });
+  });
+
+  it('loads the participants and their participation information for a year (on or after 2016)', function (done) {
+    eventStore.state.registrationEvents = [
+      events.participantWasRegistered('single', 2, 'session-id', 'memberId')
+    ];
+
+    subscriber = new Subscriber({id: 'memberId'});
+    subscriber.participationOf('2020').state.roommate = 'My buddy';
+
+    socratesWikiService.getParticipantsFor('2020', function (err, participants) {
+      expect(participants).to.have.length(1);
+      expect(participants[0].participation.roommate()).to.be('My buddy');
+      done(err);
+    });
+  });
+
+});
